Update Create_Memory.java
Method WRITE and READ from internal storage:

Method "saveToInternalStorage" converts a Bitmap to .JPG and saves it on the Internal Storage. It returns te Absolute Path (future Database).

Method "loadImageFromStorage" finds the PNG image by giving the Path as argument. Converts it to .jpg and gives it to an possible ImageView.